[MOB-1798] Bound the WAL replay by the file's size, and split the decoder for lint
